+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [[NEXT]](https://github.com/iExecBlockchainComputing/iexec-common/releases/tag/vNEXT) 2023
 
+### Quality
+- Mark classes from `com.iexec.common.sdk.order` and `com.iexec.common.sdk.order.payload` packages deprecated. (#414)
+
 ## [[8.1.0]](https://github.com/iExecBlockchainComputing/iexec-common/releases/tag/v8.1.0) 2023-05-11
 
 ### New Features